chore: move helpers
