+++ packages/ui-components/src/display/natValue/helpers/captureNum.js
